[OPP-1378] legge til sf-legacy-controller for delsvar
